Some more updates on issue #271
In GameDbSyncEvents.h, I changed my structs' m_id to uint64_t, since Entity::GetAccountId returns uint64_t as well.
And I tried to make GameDbSyncService closer to WorldSimulation

+++ Projects/CoX/Servers/GameDatabase/GameDBSyncEvents.h
@@ -251,28 +251,28 @@ ONE_WAY_MESSAGE(SetClientOptions)
 
 struct GuiUpdateData
 {
-    uint32_t m_id;
+    uint64_t m_id;
     QString m_gui;
 };
 ONE_WAY_MESSAGE(GuiUpdate)
 
 struct OptionsUpdateData
 {
-    uint32_t m_id;
+    uint64_t m_id;
     QString m_options;
 };
 ONE_WAY_MESSAGE(OptionsUpdate)
 
 struct KeybindsUpdateData
 {
-    uint32_t m_id;
+    uint64_t m_id;
     QString m_keybinds;
 };
 ONE_WAY_MESSAGE(KeybindsUpdate)
 
 struct PlayerUpdateData
 {
-    uint32_t m_id;
+    uint64_t m_id;
     QString m_player_data;
 };
 ONE_WAY_MESSAGE(PlayerUpdate)
